+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+  - Deployment - Additional filter
   - Dependency - nuvla-api-version v2.0.9
   - Sidebar item has an href
   - Deployment page - Use card component